Stock_analysis_AI_agents/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Config
pydantic-settings>=2.4.0

# LLM + Agents
langchain-core>=1.0.0
langchain-groq>=0.2.0
langchain-openai>=0.3.0
crewai>=0.80.0

# Stock data
yfinance>=0.2.40
numpy>=1.26.0
pandas>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
